+++ packages/vite/src/node/optimizer/index.ts
@@ -1,6 +1,6 @@
-import fs from 'fs'
-import path from 'path'
-import { performance } from 'perf_hooks'
+import fs from 'node:fs'
+import path from 'node:path'
+import { performance } from 'node:perf_hooks'
 import _debug from 'debug'
 import colors from 'picocolors'
 import type { BuildOptions as EsbuildBuildOptions } from 'esbuild'
@@ -534,15 +534,30 @@ export async function runOptimizeDeps(
     flatIdToExports[flatId] = exportsData
   }
 
+  // esbuild automatically replaces process.env.NODE_ENV for platform 'browser'
+  // In lib mode, we need to keep process.env.NODE_ENV untouched, so to at build
+  // time we replace it by __vite_process_env_NODE_ENV. This placeholder will be
+  // later replaced by the define plugin
+  const define = {
+    'process.env.NODE_ENV': isBuild
+      ? '__vite_process_env_NODE_ENV'
+      : JSON.stringify(process.env.NODE_ENV || config.mode)
+  }
+
   const start = performance.now()
 
   const result = await build({
     absWorkingDir: process.cwd(),
     entryPoints: Object.keys(flatIdDeps),
     bundle: true,
-    // Ensure resolution is handled by esbuildDepPlugin and
-    // avoid replacing `process.env.NODE_ENV` for 'browser'
-    platform: 'neutral',
+    // We can't use platform 'neutral', as esbuild has custom handling
+    // when the platform is 'node' or 'browser' that can't be emulated
+    // by using mainFields and conditions
+    platform:
+      config.build.ssr && config.ssr?.target !== 'webworker'
+        ? 'node'
+        : 'browser',
+    define,
     format: 'esm',
     target: isBuild ? config.build.target || undefined : ESBUILD_MODULES_TARGET,
     external: config.optimizeDeps?.exclude,