+++ packages/vite/src/node/publicUtils.ts
@@ -4,6 +4,8 @@
  * Please control the side-effects by checking the ./dist/node-cjs/publicUtils.cjs bundle
  */
 export { VERSION as version } from './constants'
+export { version as esbuildVersion } from 'esbuild'
+export { VERSION as rollupVersion } from 'rollup'
 export {
   splitVendorChunkPlugin,
   splitVendorChunk