+++ playground/tsconfig-json/__tests__/tsconfig-json.spec.ts
@@ -1,5 +1,5 @@
-import path from 'path'
-import fs from 'fs'
+import path from 'node:path'
+import fs from 'node:fs'
 import { transformWithEsbuild } from 'vite'
 import { browserLogs } from '~utils'