+++ playground/worker/__tests__/relative-base/relative-base-worker.spec.ts
@@ -1,5 +1,5 @@
-import fs from 'fs'
-import path from 'path'
+import fs from 'node:fs'
+import path from 'node:path'
 import { isBuild, page, testDir, untilUpdated } from '~utils'
 
 test('normal', async () => {